Moved SearchFilters styles into common
