--- client/src/components/common/SearchFilter/index.js
@@ -1,1 +1,0 @@
-export * from "./SearchFilter";
